requirements.txt
Django==4.2.7
djangorestframework==3.13.1
Pillow==10.1.0
requests==2.31.0
imagehash==4.3.1
python-dotenv==1.0.0
